cleanup_cloud_run/requirements.txt
# file needs to exist - requirement of Google Cloud Functions

google-api-python-client